+++ packages/session-cookie/src/session-cookie.ts
@@ -1,5 +1,5 @@
-import { Cookies } from '@ts-stack/cookies';
-import { injectable, optional, RequestContext } from '@ditsmod/core';
+import { Cookies, NodeRequest, NodeResponse } from '@ts-stack/cookies';
+import { inject, injectable, NODE_REQ, NODE_RES, optional, fromSelf } from '@ditsmod/core';
 
 import { SessionCookieOptions } from './types';
 
@@ -9,16 +9,20 @@ export class SessionCookie {
   protected cookies: Cookies;
   protected maxAge: number;
 
-  constructor(ctx: RequestContext, @optional() protected opts: SessionCookieOptions) {
+  constructor(
+    @fromSelf() @inject(NODE_REQ) nodeReq: NodeRequest,
+    @fromSelf() @inject(NODE_RES) nodeRes: NodeResponse,
+    @optional() protected opts: SessionCookieOptions
+  ) {
     this.opts = { ...(opts || {}) };
-    this.cookies = new Cookies(ctx.nodeReq, ctx.nodeRes);
+    this.cookies = new Cookies(nodeReq, nodeRes);
     this.opts.cookieName = this.opts.cookieName || 'session_id';
     this.maxAge = this.opts.maxAge === undefined ? 1000 * 60 * 60 * 24 : this.opts.maxAge; // By default - 24 hours
 
-    const writeHead = ctx.nodeRes.writeHead as Function;
-    ctx.nodeRes.writeHead = (...args: any[]) => {
+    const writeHead = nodeRes.writeHead as Function;
+    nodeRes.writeHead = (...args: any[]) => {
       this.updateSessionCookie();
-      return writeHead.apply(ctx.nodeRes, args);
+      return writeHead.apply(nodeRes, args);
     };
   }